file_storage/requirements.txt
asgiref==3.8.1
attrs==25.3.0
Django==5.2.3
djangorestframework==3.16.0
dnspython==2.7.0
dotenv==0.9.9
drf-spectacular==0.28.0
inflection==0.5.1
isort==6.0.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
pymongo==4.13.2
python-dotenv==1.1.1
python-json-logger==3.3.0
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.26.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
